layers: Fix heap mapping bindingCount overflow

+++ layers/utils/descriptor_utils.cpp
@@ -108,7 +108,8 @@ bool IsResourceVaribleInMapping(const VkDescriptorSetAndBindingMappingEXT& mappi
                                 const spirv::ResourceInterfaceVariable& resource_variable) {
     const uint32_t descriptor_set = resource_variable.decorations.set;
     const uint32_t descriptor_binding = resource_variable.decorations.binding;
-    const uint32_t last_binding = mapping.firstBinding + mapping.bindingCount;
+    // bindingCount could be UINT32_MAX
+    const uint64_t last_binding = mapping.firstBinding + uint64_t(mapping.bindingCount);
     return (mapping.descriptorSet == descriptor_set && descriptor_binding >= mapping.firstBinding &&
             descriptor_binding < last_binding && ResourceTypeMatchesBinding(mapping.resourceMask, resource_variable));
 }
